+++ examples/demo/run_server.py
@@ -11,8 +11,8 @@ server = sdk.Server(14514, 14515, {
     'another_client': 'another_team'
 })
 
-async def general_data_callback(client_name: str, title: str, data: bytes):
-    print(f'general_data_callback: {client_name} {title} {data}')
+async def topic_message_callback(client_name: str, title: str, data: bytes):
+    print(f'topic_message_callback: {client_name} {title} {data}')
 
 async def robot_control_callback(client_name: str, rb_ctrl: sdk.RobotControl):
         if rb_ctrl.head is not None:
@@ -36,7 +36,7 @@ async def main():
         },
         simulation_rate=1.3))
     
-    await server.register_general_data_callback(general_data_callback)
+    await server.register_topic_message_callback(topic_message_callback)
     await server.register_robot_control_callback(robot_control_callback)
 
     for i in range(600):
@@ -49,7 +49,7 @@ async def main():
                 game_info.score['another_team'] += random.randint(0, 10)
                 game_info.simulation_rate = random.uniform(0.5, 2.0)
 
-            await server.push_general_data('example_client', 'server/title', f'data/{random.randint(0, 1000000)}'.encode())
+            await server.push_topic_message('example_client', 'server/title', f'data/{random.randint(0, 1000000)}'.encode())
 
 
         if i % 1 == 0: