+++ src/soccerxcomm/server.py
@@ -44,7 +44,7 @@ class Server:
         self._client_team_map: Dict[str, str] = client_team_map
         self._is_callback_registered: bool = False
         self._robot_control_callback_list: List[Callable[[str, RobotControl], Coroutine[Any, Any, None]]] = []
-        self._general_data_callback_list: List[Callable[[str, str, bytes], Coroutine[Any, Any, None]]] = []
+        self._topic_message_callback_list: List[Callable[[str, str, bytes], Coroutine[Any, Any, None]]] = []
 
         # Components
         self._controller_network_server: INetworkServer = HttpServer(
@@ -104,19 +104,19 @@ class Server:
             'shape': list(image.shape),
         }), token)
 
-    async def push_general_data(self, token: str, title: str, data: bytes) -> None:
-        """Pushes the general data to the client.
+    async def push_topic_message(self, token: str, topic: str, data: bytes) -> None:
+        """Pushes the topic message to the client.
 
         Args:
             token: The token of the client.
-            title: The title of the data.
-            data: The data.
+            topic: The topic of the message.
+            data: The data bytes of the message.
         """
 
         await self._controller_network_server.send(Message({
-            'type': 'push_general_data',
+            'type': 'push_topic_message',
             'bound_to': 'client',
-            'title': title,
+            'topic': topic,
             'data': data
         }), token)
 
@@ -155,14 +155,14 @@ class Server:
             'team': robot_status.team
         }), token)
 
-    async def register_general_data_callback(self, callback: Callable[[str, str, bytes], Coroutine[Any, Any, None]]) -> None:
-        """Registers a callback for the general data.
+    async def register_topic_message_callback(self, callback: Callable[[str, str, bytes], Coroutine[Any, Any, None]]) -> None:
+        """Registers a callback for topic messages.
 
         Args:
             callback: The callback.
         """
 
-        self._general_data_callback_list.append(callback)
+        self._topic_message_callback_list.append(callback)
 
     async def register_robot_control_callback(self, callback: Callable[[str, RobotControl], Coroutine[Any, Any, None]]) -> None:
         """Registers a callback for the robot control.
@@ -233,11 +233,11 @@ class Server:
                 for callback in self._robot_control_callback_list:
                     await callback(client_token, RobotControl(head, movement, kick))
 
-            elif message_type == 'push_general_data':
+            elif message_type == 'push_topic_message':
                 obj = message.to_dict()
 
-                for callback in self._general_data_callback_list:
-                    await callback(client_token, obj['title'], obj['data'])
+                for callback in self._topic_message_callback_list:
+                    await callback(client_token, obj['topic'], obj['data'])
 
         except Exception as e:
             self._logger.error(f"Failed to handle message: {e}")