+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- General data transmission.
+- Topic message transmission.
 
 ## [0.2.0] - 2023-08-20